user registration, redirect, save a new user & basic form validation
