+++ static/style.css
@@ -61,9 +61,18 @@ img {
   background: var(--secondary-err)
 }
 
+.message_success {
+  text-align: center;
+  padding: 5px;
+  border: #ccc dotted 2px;
+  margin-top: 2rem;
+  color: var(--primary-text);
+  background: var(--primary-btn);
+}
+
 /* loading  */
 div#loading {
-  top: 30rem;
+  top: 15rem;
   position: absolute;
   z-index: 1000;
   height: 24px;
