+++ templates/login.html
@@ -4,6 +4,15 @@
 </title>
 {% endblock %} {%block content%}
 <div class="container">
+
+    <!-- {% if message %}
+    <p class="message">{{ message | safe }}</p>
+    {% endif %} -->
+
+    {% with messages = get_flashed_messages() %} {% if messages %} {%for message in messages %}
+    <p class="message_success">{{message}}</p>
+    {% endfor %} {% endif %} {% endwith %}
+
     <div class="row">
         <div class="col-12 mx-auto">
 
@@ -13,16 +22,19 @@
                     <h3>Login</h3>
                 </div>
                 <div class="card-body">
-                    <form action="#">
+                    <form action="POST">
                         <div class="form-group">
-                            <h6>Email:</h6>
-                            <input type="text" name="email" placeholder="johndoe@test.com" />
+                            {{form.hidden_tag()}} {{form.email.label}}{{form.email()}}
+                            <!-- <h6>Email:</h6> -->
+                            <!-- <input type="text" name="email" placeholder="johndoe@test.com" /> -->
                         </div>
                         <div class="form-group">
-                            <h6>Password:</h6>
-                            <input type="password" name="password" />
+                            {{form.password.label}}{{form.password()}}
+                            <!-- <h6>Password:</h6>
+                            <input type="password" name="password" /> -->
                         </div>
-                        <button id="register" type="submit" class="btn btn-primary">Submit</button>
+                        {{form.submit()}}
+                        <!-- <button id="register" type="submit" class="btn btn-primary">Submit</button> -->
                     </form>
                 </div>
             </div>
