+++ templates/register.html
@@ -5,6 +5,11 @@
 {% endblock %} {%block content%}
 
 <div class="container">
+
+    {% if message %}
+    <p class="message">{{ message | safe }}</p>
+    {% endif %}
+
     <div class="row text-center">
         <div class="col-12 mx-auto">
 
@@ -26,9 +31,16 @@
                             <h6>Password:</h6>
                             <input type="password" name="password" />
                         </div>
+                        <div class="form-group">
+                            <h6>Confirm Password:</h6>
+                            <input type="password" name="confirm_pw" />
+                        </div>
 
-                        <button id="register" type="submit" class="btn btn-primary">Submit</button>
-
+                        <!-- <button id="register" type="submit" class="btn btn-primary">Submit</button> -->
+                        <div id="loading" style="display:none;">
+                            <img src="../static/loading.gif" alt="loading spinner" />
+                        </div>
+                        <input onclick="$('#loading').show();" type="submit" value="Submit" class="btn" />
                     </form>
                 </div>
             </div>
